+++ lib/Api/GeneralMerchantsApi.php
@@ -308,12 +308,13 @@ class GeneralMerchantsApi
      * Read General Merchant
      *
      * @param string $general_merchant_id General Merchant ID (required)
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return \Secuconnect\Client\Model\GeneralMerchantsProductModel
      */
-    public function getOne($general_merchant_id)
+    public function getOne($general_merchant_id, $expand = null)
     {
-        list($response) = $this->getOneWithHttpInfo($general_merchant_id);
+        list($response) = $this->getOneWithHttpInfo($general_merchant_id, $expand);
         return $response;
     }
 
@@ -323,10 +324,11 @@ class GeneralMerchantsApi
      * Read General Merchant
      *
      * @param string $general_merchant_id General Merchant ID (required)
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return array of \Secuconnect\Client\Model\GeneralMerchantsProductModel, HTTP status code, HTTP response headers (array of strings)
      */
-    public function getOneWithHttpInfo($general_merchant_id)
+    public function getOneWithHttpInfo($general_merchant_id, $expand = null)
     {
         // verify the required parameter 'general_merchant_id' is set
         if ($general_merchant_id === null || (is_array($general_merchant_id) && count($general_merchant_id) === 0)) {
@@ -346,6 +348,10 @@ class GeneralMerchantsApi
         }
         $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
 
+        // query params
+        if ($expand !== null) {
+            $queryParams['expand'] = $this->apiClient->getSerializer()->toQueryValue($expand);
+        }
         // path params
         if ($general_merchant_id !== null) {
             $resourcePath = str_replace(