+++ lib/Api/PaymentTransactionsApi.php
@@ -834,12 +834,13 @@ class PaymentTransactionsApi
      * Read Payment Transaction
      *
      * @param string $payment_transaction_id Payment Transaction ID (required)
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return \Secuconnect\Client\Model\PaymentTransactionsProductModel
      */
-    public function getOne($payment_transaction_id)
+    public function getOne($payment_transaction_id, $expand = null)
     {
-        list($response) = $this->getOneWithHttpInfo($payment_transaction_id);
+        list($response) = $this->getOneWithHttpInfo($payment_transaction_id, $expand);
         return $response;
     }
 
@@ -849,10 +850,11 @@ class PaymentTransactionsApi
      * Read Payment Transaction
      *
      * @param string $payment_transaction_id Payment Transaction ID (required)
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return array of \Secuconnect\Client\Model\PaymentTransactionsProductModel, HTTP status code, HTTP response headers (array of strings)
      */
-    public function getOneWithHttpInfo($payment_transaction_id)
+    public function getOneWithHttpInfo($payment_transaction_id, $expand = null)
     {
         // verify the required parameter 'payment_transaction_id' is set
         if ($payment_transaction_id === null || (is_array($payment_transaction_id) && count($payment_transaction_id) === 0)) {
@@ -872,6 +874,10 @@ class PaymentTransactionsApi
         }
         $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
 
+        // query params
+        if ($expand !== null) {
+            $queryParams['expand'] = $this->apiClient->getSerializer()->toQueryValue($expand);
+        }
         // path params
         if ($payment_transaction_id !== null) {
             $resourcePath = str_replace(