+++ lib/Api/PrepaidSalesApi.php
@@ -416,12 +416,13 @@ class PrepaidSalesApi
      * Read Prepaid Sale
      *
      * @param string $prepaid_sale_id Prepaid sale ID (required)
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return \Secuconnect\Client\Model\PrepaidSalesProductModel
      */
-    public function getOne($prepaid_sale_id)
+    public function getOne($prepaid_sale_id, $expand = null)
     {
-        list($response) = $this->getOneWithHttpInfo($prepaid_sale_id);
+        list($response) = $this->getOneWithHttpInfo($prepaid_sale_id, $expand);
         return $response;
     }
 
@@ -431,10 +432,11 @@ class PrepaidSalesApi
      * Read Prepaid Sale
      *
      * @param string $prepaid_sale_id Prepaid sale ID (required)
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return array of \Secuconnect\Client\Model\PrepaidSalesProductModel, HTTP status code, HTTP response headers (array of strings)
      */
-    public function getOneWithHttpInfo($prepaid_sale_id)
+    public function getOneWithHttpInfo($prepaid_sale_id, $expand = null)
     {
         // verify the required parameter 'prepaid_sale_id' is set
         if ($prepaid_sale_id === null || (is_array($prepaid_sale_id) && count($prepaid_sale_id) === 0)) {
@@ -454,6 +456,10 @@ class PrepaidSalesApi
         }
         $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
 
+        // query params
+        if ($expand !== null) {
+            $queryParams['expand'] = $this->apiClient->getSerializer()->toQueryValue($expand);
+        }
         // path params
         if ($prepaid_sale_id !== null) {
             $resourcePath = str_replace(